+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## v1.15.0
 
-*2025-08-12*
+*2025-08-14*
 
 **Added**
 
 - New refactoring action provider for Wikitext
+- Fix-all actions for WikiLint
 
 ## v1.14.1